Add ability to retrieve and store state of filter form from ViewModel
Comments:
Holy shit this actually took me a month of on-and-off coding to get
it to work.
This is actually the single hardest feature I have ever done, and
hopefully it gets easier from here.
It's actually not fully complete but the core functionality is there,
which is what matters most.

Changelist:
+ Make WorkFilterFragment access the WorkFilterRequest in the
  WorkListViewModel directly to retrieve filter state.
+ Make WorkFilterFragment store state through the WorkListViewModel's
  method.

Todo:
~ Add the remaining fields into the form
~ Find a way to handle dates
~ Convert the layout to a ConstraintLayout? (this is not critical)